#12 [REFACTOR] 모든 알림 가져오기 기능을 NotificationService로 이동
- NotificationEventServce: fcm 알림 메시지를 전송하는 Service
- NotificationService: 알림 관련 기능을 처리하는 Service
